Fixed and added new Services
