+++ src/services/spreadsheet.service.ts
@@ -1,35 +1,14 @@
-import { Request, Response } from "express";
-import { connect } from "../database.mysql/database";
-import { ParticipantBatchMap } from "../interface/taskbotic";
+import { Request, Response } from "express";  
 import { GoogleSpreadsheet } from "google-spreadsheet"
 import { creds } from "../resources/credentials/credentials";
-const otpGenerator = require('otp-generator') as any;
+const otpGenerator = require('otp-generator') as any; 
 
 let doc = new GoogleSpreadsheet(
     "1J1U8awhcpyYrQg5z4Vx9uTu8Jn3-ycwZTGnZCpOPSMo"
 );
 
-export class TaskBoticService {
-    constructor() { }
-
-    async getParticipantsBatchMapping(req: Request, res: Response) {
-        try {
-            let db: any = req.headers.db;
-            const conn = await connect(db);
-            let selectQuery:any =`select * from participants_batch_map`
-            const participants = await conn.query(selectQuery);
-            res.status(200).json(
-                participants[0]
-            );
-            return conn.end();
-        }
-        catch (err) {
-            res.status(500).json({
-                message: err
-            })
-        }
-    }
-    
+export class SpreadsheetService{ 
+    constructor(){} 
     async getSpreadsheetData(req: Request, res: Response) {
         try{
             await doc.useServiceAccountAuth({
